requirements.txt
matplotlib==3.7.1
numpy==1.24.3
scipy==1.10.1
bitarray==2.7.4
torch-hd==5.1.2
torch==2.0.1
pillow==9.5.0